background/requirements.txt
#https://www.lfd.uci.edu/~gohlke/pythonlibs/#wordcloud
mysqlclient==2.0.3
djangorestframework==3.12.2